Add article card, fix posts page, update insights page
