+++ src/app/posts/[id]/post.module.scss
