--- src/sections/insights/hero/hero.module.scss
@@ -1,100 +1,0 @@
-@use "styles/global" as *;
-
-.root {
-  background: conic-gradient(
-    from 270deg at 50% 50%,
-    rgba(207, 236, 87, 0.1) -80deg,
-    rgba(22, 29, 37, 0.2) 43deg,
-    rgba(207, 236, 87, 0.5) 290deg,
-    #161d25 280deg,
-    rgba(207, 236, 87, 0.5) 280deg,
-    #161d25 280.26deg,
-    rgba(207, 236, 87, 0.1) 280deg,
-    rgba(22, 29, 37, 0.2) 400deg
-  );
-
-  background-color: black;
-  padding-top: 236px;
-  padding-bottom: 236px;
-  color: white;
-  position: relative;
-
-  margin-bottom: 140px;
-}
-
-.container {
-  margin-left: 160px;
-}
-
-.title {
-  @include fontJetBrains;
-  text-transform: uppercase;
-  font-size: 120px;
-  font-weight: 300;
-  line-height: 100%;
-  letter-spacing: -0.08em;
-
-  max-width: 949px;
-  margin-bottom: 48px;
-
-  span {
-    background:
-      linear-gradient(
-        90deg,
-        rgba(207, 235, 87, 0.4) 0%,
-        rgba(207, 235, 87, 0) 100%
-      ),
-      #ffffff;
-    -webkit-background-clip: text;
-    background-clip: text;
-    -webkit-text-fill-color: transparent;
-    --text-fill-color: transparent;
-  }
-}
-
-.text {
-  @include fontAbel;
-  font-size: 18px;
-  font-weight: 400;
-  line-height: 140%;
-  color: rgba(255, 255, 255, 0.6);
-
-  margin-bottom: 48px;
-
-  max-width: 380px;
-}
-
-.button {
-  @include fontJetBrains;
-  font-weight: 200;
-  font-size: 18px;
-  line-height: 100%;
-  letter-spacing: -0.08em;
-  text-transform: uppercase;
-  color: #000;
-
-  min-height: 60px;
-  padding: 30px 88px;
-  border-radius: 8px;
-  background:
-    conic-gradient(from 25deg at 14% 50%, #e2ff66 0deg, #8da720 360deg), #ffffff;
-  clip-path: polygon(
-    0 0,
-    100% 0,
-    100% calc(100% - 30px),
-    calc(100% - 30px) 100%,
-    0 100%
-  );
-
-  cursor: pointer;
-}
-
-.heroImage {
-  position: absolute;
-  bottom: 0;
-  right: 2%;
-  z-index: 0;
-
-  width: 50%;
-  height: auto;
-}
