+++ src/shared/components/cards/blog-card/blog-card.module.scss
@@ -1,0 +1,75 @@
+@use "styles/global" as *;
+
+.card {
+  // box-sizing: border-box;
+  display: flex;
+  gap: 30px;
+  padding: 16px;
+  border: 1px solid rgba(255, 255, 255, 0.16);
+
+  &__image {
+    object-fit: cover;
+    width: 50%;
+    height: auto;
+  }
+
+  &__info {
+    display: flex;
+    flex-direction: column;
+    gap: 32px;
+    padding: 10px 0;
+  }
+
+  &__title {
+    font-weight: 200;
+    font-size: 20px;
+    line-height: 90%;
+    text-transform: uppercase;
+
+    margin-bottom: 8px;
+  }
+
+  &__subtitle {
+    font-weight: 200;
+    font-size: 16px;
+    line-height: 90%;
+    text-transform: uppercase;
+    color: rgba(255, 255, 255, 0.4);
+  }
+
+  &__description {
+    @include fontAbel;
+    flex: 1;
+    font-weight: 400;
+    font-size: 16px;
+    line-height: 140%;
+    color: rgba(255, 255, 255, 0.6);
+
+    max-width: 325px;
+  }
+
+  &__btn {
+    box-sizing: border-box;
+    width: 166px;
+    padding: 24px;
+
+    background-color: black;
+    border: 1px solid rgba(255, 255, 255, 0.16);
+    border-radius: 8px;
+    clip-path: polygon(
+      0 0,
+      100% 0,
+      100% calc(100% - 30px),
+      calc(100% - 30px) 100%,
+      0 100%
+    );
+
+    @include fontJetBrains;
+    font-weight: 200;
+    font-size: 18px;
+    line-height: 100%;
+    letter-spacing: -0.08em;
+    text-transform: uppercase;
+    color: #ffffff;
+  }
+}
