+++ src/shared/components/cards/index.ts
@@ -1,0 +1,1 @@
+export { default as BlogCard } from "./blog-card/blog-card";
