+++ src/shared/constants/hero-backgrounds.ts
@@ -31,6 +31,7 @@ export const BACKGROUND_PRESETS = {
 		rgba(226, 255, 102, 0.5) 0%,
 		rgba(226, 255, 102, 0) 50%
 	)`,
+
   green: `conic-gradient(
     from 280deg at 50% 50%,
     rgba(207, 236, 87, 0.1) -80deg,
