+++ src/shared/texts/insights/index.json
@@ -1,0 +1,7 @@
+{
+  "hero": {
+    "title": "soft does Insights",
+    "tagText": "We transform market insights into globally beloved apps that power your business",
+    "btnText": "Contact us"
+  }
+}
